Adds property "healthy" to connection pool.
As mentioned in the [changelog](https://github.com/tediousjs/node-mssql#5x-to-6x-changes) between `mssql` 5.x to 6.x, the `ConnectionPool` class now has a property, `healthy` which is publicly accessible. This property is not included on the typedef. 

A fix for this issue would be to include modify the `ConnectionPool` type in @types/mssql to include the following:

```
export declare class ConnectionPool extends events.EventEmitter {
    public connected: boolean;
    ...
    public healthy: boolean; <-- New property
}
```

The corresponding code in the `mssql` library can be found [here](https://github.com/tediousjs/node-mssql/blob/master/lib/base/connection-pool.js#L81).